requirements.txt
altgraph==0.17
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.10
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
django-pdfkit==0.3.1
filelock==3.4.0
fpdf==1.7.2
fpdf2==2.5.0
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
netcat==0.0.1a0
pdfkit==0.6.0
pefile==2021.5.24
Pillow==9.0.0
pipenv==2021.11.23
platformdirs==2.4.0
psycopg2==2.9.3
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
python-dateutil==1.5
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stdiomask==0.0.6
urllib3==1.26.8
virtualenv==20.10.0
virtualenv-clone==0.5.7
whitenoise==5.3.0
